separated tuple and tuple comparator classes into separate files and added add course method
