+++ coursemaker/src/Tuple.java
@@ -1,0 +1,18 @@
+class Tuple<A, B> {
+//    class for tuple data constructor
+    private A first;
+    private B second;
+
+    public Tuple(A first, B second) {
+        this.first = first;
+        this.second = second;
+    }
+
+    public A getFirst() {
+        return first;
+    }
+
+    public B getSecond() {
+        return second;
+    }
+}
